stop steps 1 to 3 widening the cube just to reduce it
Three places held a float64 copy of something cube-sized for no reason the
answer depends on.

step1 collapsed the whole cube in one nanmean, which copies its input to
replace the NaNs; it now takes a band of image rows at a time. The split is
spatial, so each pixel still accumulates along the whole wavelength axis in
one call and the summation order is untouched.

step2 read each wavelength chunk as float64 when the cube is float32. The
chunk stays float32 and only one source's spaxels are widened at a time,
before the sum rather than during it -- asking sum() for a float64
accumulator over float32 input gives the same answer only for as long as it
keeps blocking the reduction the way it does now, which is not a promise it
makes.

step3 built the blank residual twice, once by subtracting and once by
np.where, and then made a third full copy inside nan_to_num before narrowing
to float32. The residual is now written once and narrowed a block at a time.
The order of nan_to_num and the cast cannot be flipped: on float64 an
infinity becomes ~1.8e308 and overflows back to inf on the cast, while
narrowing first turns it into ~3.4e38, a finite number the decomposition
would treat as real data.

Products are unchanged, byte for byte, on p05 and p14 -- 26 .npy files and
both whitelight.fits including their headers. Peak heap: step1 2.34 -> 0.30
GiB, step2 0.57 -> 0.34, step3 4.87 -> 3.82 (p14). No step got slower.

+++ src/skymodel/step2_object_spectra.py
@@ -81,16 +81,28 @@ def sum_spectra_by_id(cube_path, seg, ids, chunk=200, var_path=None):
         nspax = np.zeros((len(ids), nz))
 
         for j in range(0, nz, chunk):
-            d = np.asarray(hdul["DATA"].data[j:j+chunk], np.float64).reshape(-1, seg_flat.size)
-            v = np.asarray(vdul["STAT"].data[j:j+chunk], np.float64).reshape(-1, seg_flat.size)
+            # Kept at the cube's own float32: the widening is only needed for the
+            # summation below, which asks for float64 accumulation itself, so doing
+            # it here would carry a double-width copy of the chunk for nothing.
+            d = np.asarray(hdul["DATA"].data[j:j+chunk], np.float32).reshape(-1, seg_flat.size)
+            v = np.asarray(vdul["STAT"].data[j:j+chunk], np.float32).reshape(-1, seg_flat.size)
 
             ok = np.isfinite(d) & np.isfinite(v) & (v > 0)
             d  = np.where(ok, d, 0.0)
             v  = np.where(ok, v, 0.0)
 
             for k, idx in enumerate(members):
-                flux[k,  j:j+chunk] = d[:, idx].sum(axis=1)
-                var[k,   j:j+chunk] = v[:, idx].sum(axis=1)
+                # Widened before the sum, not during it: a source can cover
+                # thousands of spaxels, and accumulating that many float32 terms
+                # at float32 width loses digits. Asking sum() for a float64
+                # accumulator over float32 input would give the same answer only
+                # for as long as it keeps blocking the reduction the way it does
+                # now, which is not a promise it makes; widening first is the
+                # same arithmetic by construction. Only one source's spaxels are
+                # widened at a time, so the chunk itself stays float32. nspax
+                # counts a boolean and keeps the integer accumulator sum picks.
+                flux[k,  j:j+chunk] = d[:, idx].astype(np.float64).sum(axis=1)
+                var[k,   j:j+chunk] = v[:, idx].astype(np.float64).sum(axis=1)
                 nspax[k, j:j+chunk] = ok[:, idx].sum(axis=1)
 
     return flux, var, nspax
